+++ lib/themes/dark_theme.dart
@@ -5,7 +5,10 @@ import 'package:game_list/themes/nord.dart';
 ThemeData darkTheme = ThemeData(
   fontFamily: 'Inter',
   accentColor: Nord.nord11,
-  appBarTheme: AppBarTheme(color: Nord.nord0),
+  appBarTheme: AppBarTheme(
+    color: Nord.nord0,
+    brightness: Brightness.dark,
+  ),
   bottomNavigationBarTheme: BottomNavigationBarThemeData(
       backgroundColor: Nord.nord0,
       selectedItemColor: Nord.nord11,